// src/app/shared/model/stockmaster.ts
export class StockMaster {

    SpDocNo: string;
    SpDocSrNo: number;
    StockCode: string;
    StockDesc: string;
    StockOrigin: string;
    StockCartQty: number;
    StockRate: number;
    StockUnits: string;
    DisplayRate: number;
    DisplayUnits: string;
    StockDetail1: string;
    StockAmount: number;
    StockImgLink: string;
    StockFavourites: string;
    StockOldPrice: number;
    DiscountedPrice: number;
    RibbonText: string;
    RibbonColor: string;
    category_icon: string;
    featured_icon: string;
    Meat_icon: string;
    FamilyDesc: string;
    isPromotionText: string;
    StockCounter: number;
    ExtraInfo: string;
    StockCounter2: number;
    OutStock: string;
    lblGray4: string;
    DepositRate: number;
    DepositMessage: string;

 }

 export class AppMenus {
    ActionType: string;
    BackCircle: string;
    BackColor: string;
    ClickAction: string;
    Competition_Code: string;
    DocDesc: string;
    DocNo: string;
    img_link: string;
    small_icon: string;
    Sorting: string;
 }